LayoutTests/inspector/model/remote-object-get-properties-expected.txt


-----------------------------------------------------
EXPRESSION: window.simpleObject
type: object
description: Object

OWN PROPERTIES:
    a
    b
    __proto__

DISPLAYABLE PROPERTIES:
    a
    b
    __proto__

ALL PROPERTIES:
    a
    b
    toString
    toLocaleString
    valueOf
    hasOwnProperty
    propertyIsEnumerable
    isPrototypeOf
    __defineGetter__
    __defineSetter__
    __lookupGetter__
    __lookupSetter__
    constructor
    __proto__
-----------------------------------------------------

-----------------------------------------------------
EXPRESSION: window.loadEvent
type: object
description: Event

OWN PROPERTIES:
    isTrusted
    __proto__

DISPLAYABLE PROPERTIES:
    isTrusted
    type
    target
    currentTarget
    eventPhase
    bubbles
    cancelable
    timeStamp
    defaultPrevented
    srcElement
    returnValue
    cancelBubble
    clipboardData
    __proto__

ALL PROPERTIES:
    isTrusted
    constructor
    type
    target
    currentTarget
    eventPhase
    bubbles
    cancelable
    timeStamp
    defaultPrevented
    srcElement
    returnValue
    cancelBubble
    clipboardData
    stopPropagation
    preventDefault
    initEvent
    stopImmediatePropagation
    NONE
    CAPTURING_PHASE
    AT_TARGET
    BUBBLING_PHASE
    MOUSEDOWN
    MOUSEUP
    MOUSEOVER
    MOUSEOUT
    MOUSEMOVE
    MOUSEDRAG
    CLICK
    DBLCLICK
    KEYDOWN
    KEYUP
    KEYPRESS
    DRAGDROP
    FOCUS
    BLUR
    SELECT
    CHANGE
    toString
    toLocaleString
    valueOf
    hasOwnProperty
    propertyIsEnumerable
    isPrototypeOf
    __defineGetter__
    __defineSetter__
    __lookupGetter__
    __lookupSetter__
    __proto__
-----------------------------------------------------

-----------------------------------------------------
EXPRESSION: window.complexObject
type: object
description: SuperFoo

OWN PROPERTIES:
    _bar
    _foo
    __proto__

DISPLAYABLE PROPERTIES:
    _bar
    _foo
    __proto__

ALL PROPERTIES:
    _bar
    _foo
    constructor
    getterProperty
    foo
    BAR_CONSTANT
    bar
    setterOnly
    toString
    toLocaleString
    valueOf
    hasOwnProperty
    propertyIsEnumerable
    isPrototypeOf
    __defineGetter__
    __defineSetter__
    __lookupGetter__
    __lookupSetter__
    __proto__
-----------------------------------------------------

-----------------------------------------------------
EXPRESSION: window.badGetterObject
type: object
description: ClassWithBadGetter

OWN PROPERTIES:
    __proto__

DISPLAYABLE PROPERTIES:
    __proto__

ALL PROPERTIES:
    constructor
    badGetter
    toString
    toLocaleString
    valueOf
    hasOwnProperty
    propertyIsEnumerable
    isPrototypeOf
    __defineGetter__
    __defineSetter__
    __lookupGetter__
    __lookupSetter__
    __proto__
-----------------------------------------------------

-----------------------------------------------------
EXPRESSION: window.unboundFunction
type: function
description: function () { console.log(arguments); }

OWN PROPERTIES:
    arguments
    caller
    length
    name
    prototype
    __proto__

DISPLAYABLE PROPERTIES:
    arguments
    caller
    length
    name
    prototype
    __proto__

ALL PROPERTIES:
    arguments
    caller
    length
    name
    prototype
    toString
    apply
    call
    bind
    constructor
    Symbol(Symbol.hasInstance)
    toLocaleString
    valueOf
    hasOwnProperty
    propertyIsEnumerable
    isPrototypeOf
    __defineGetter__
    __defineSetter__
    __lookupGetter__
    __lookupSetter__
    __proto__
-----------------------------------------------------

-----------------------------------------------------
EXPRESSION: window.boundFunction
type: function
description: function () {
    [native code]
}

OWN PROPERTIES:
    name
    length
    arguments
    caller
    __proto__
    targetFunction
    boundThis
    boundArgs

DISPLAYABLE PROPERTIES:
    name
    length
    arguments
    caller
    __proto__
    targetFunction
    boundThis
    boundArgs

ALL PROPERTIES:
    name
    length
    arguments
    caller
    toString
    apply
    call
    bind
    constructor
    Symbol(Symbol.hasInstance)
    toLocaleString
    valueOf
    hasOwnProperty
    propertyIsEnumerable
    isPrototypeOf
    __defineGetter__
    __defineSetter__
    __lookupGetter__
    __lookupSetter__
    __proto__
    targetFunction
    boundThis
    boundArgs
-----------------------------------------------------

-----------------------------------------------------
EXPRESSION: window.objectWithSymbolProperties
type: object
description: Object

OWN PROPERTIES:
    prop
    prop2
    Symbol()
    Symbol(sym)
    Symbol(sym)
    Symbol()
    __proto__

DISPLAYABLE PROPERTIES:
    prop
    prop2
    Symbol()
    Symbol(sym)
    Symbol(sym)
    Symbol()
    __proto__

ALL PROPERTIES:
    prop
    prop2
    Symbol()
    Symbol(sym)
    Symbol(sym)
    Symbol()
    toString
    toLocaleString
    valueOf
    hasOwnProperty
    propertyIsEnumerable
    isPrototypeOf
    __defineGetter__
    __defineSetter__
    __lookupGetter__
    __lookupSetter__
    constructor
    __proto__
-----------------------------------------------------
DONE
